get_offset_in_file: raise KeyError if not found
Raise an exception to make sure that we always consider this condition
an error.

Note that an offset of 'None' is not necessarily an error, e.g. when
doing a move refactor, an offset of 'None' would mean move the whole
module. This is surely not what someone would mean if they asked to move
'MyClasss' (mis-spelled).